* make the search filter its own component and file * add species tree poc
