+++ app/src/features/home/plan/PlanPage.tsx
@@ -23,6 +23,7 @@ import { DatabaseContext } from 'contexts/DatabaseContext';
 import React, { useContext, useEffect, useState } from 'react';
 import KMLUpload from 'components/map-buddy-components/KMLUpload';
 import { ExpandMore } from '@material-ui/icons';
+import ActivityDataFilter from 'components/activities-search-controls/ActivitiesFilter';
 
 interface IPlanPageProps {
   classes?: any;
@@ -91,159 +92,6 @@ const useStyles = makeStyles((theme) => ({
   }
 }));
 
-interface IActivityChoices {
-  activityType: string;
-  includePhotos: boolean;
-  includeForms: boolean;
-}
-
-const ActivityDataToCacheChooser: React.FC<any> = (props) => {
-  const databaseContext = useContext(DatabaseContext);
-  //todo db persist, ressurect, & update
-  const [activityChoices, setActivityChoices] = useState([]);
-
-  const saveChoices = async () => {
-    // this is what fixed the main map
-    await databaseContext.database.upsert('trip', (tripDoc) => {
-      return { ...activityChoices };
-    });
-  };
-
-  /*
-  const [doc, setDoc] = useState(null);
-
-  useEffect(() => {
-    const getPreviousTripOptions = async () => {
-      const appState = await databaseContext.database.find({ selector: { _id: 'AppState' } });
-
-      if (!appState || !appState.docs || !appState.docs.length) {
-        return;
-      }
-
-      const doc = await databaseContext.database.find({ selector: { _id: appState.docs[0].activeActivity } });
-
-      setDoc(doc.docs[0]);
-    };
-
-    getActivityData();
-  }, [databaseContext]);
-
-  if (!doc) {
-    return <CircularProgress />;
-  }
-  */
-  useEffect(() => {
-    saveChoices();
-  }, [activityChoices]);
-
-  const addActivityChoice = (newActivity: IActivityChoices) => {
-    setActivityChoices([...activityChoices, newActivity]);
-  };
-
-  const updateActivityChoice = (updatedActivity: IActivityChoices, index: number) => {
-    let updatedActivityChoices = activityChoices;
-    updatedActivityChoices[index] = updatedActivity;
-    setActivityChoices([...updatedActivityChoices]);
-  };
-
-  const deleteActivityChoice = (index: number) => {
-    let copy = [...activityChoices];
-    copy.splice(index, 1);
-    setActivityChoices(copy);
-  };
-
-  const classes = useStyles();
-
-  return (
-    <>
-      <Button
-        className={classes.activityRecordPickerAddButton}
-        onClick={() => {
-          addActivityChoice({
-            activityType: 'Observation',
-            includePhotos: false,
-            includeForms: false
-          });
-        }}>
-        add new
-      </Button>
-      <List>
-        {activityChoices.map((activityChoice, index) => {
-          return (
-            <ListItem key={index}>
-              <Paper elevation={5} className={classes.activityRecordQueryParmsRow}>
-                <Grid container spacing={3}>
-                  <Grid item xs={3}>
-                    <InputLabel id="demo-simple-select-label">Activity Type</InputLabel>
-                    <Select
-                      labelId="demo-simple-select-label"
-                      id="demo-simple-select"
-                      value={activityChoices[index].activityType}
-                      defaultValue="Select an Activity Type"
-                      onChange={(e) => {
-                        updateActivityChoice(
-                          {
-                            ...activityChoices[index],
-                            activityType: e.target.value
-                          },
-                          index
-                        );
-                      }}>
-                      <MenuItem value={'Observation'}>Observation</MenuItem>
-                      <MenuItem value={'Treatment'}>Treatment</MenuItem>
-                      <MenuItem value={'Monitoring'}>Monitoring</MenuItem>
-                    </Select>
-                  </Grid>
-                  <Grid item xs={3}>
-                    <InputLabel>Photos</InputLabel>
-                    <Switch
-                      checked={activityChoices[index].includePhotos}
-                      onChange={(e) => {
-                        updateActivityChoice(
-                          {
-                            ...activityChoices[index],
-                            includePhotos: !activityChoices[index].includePhotos
-                          },
-                          index
-                        );
-                      }}
-                    />
-                  </Grid>
-                  <Grid item xs={3}>
-                    <InputLabel>Forms</InputLabel>
-                    <Switch
-                      checked={activityChoices[index].includeForms}
-                      onChange={(e) => {
-                        updateActivityChoice(
-                          {
-                            ...activityChoices[index],
-                            includeForms: !activityChoices[index].includeForms
-                          },
-                          index
-                        );
-                      }}
-                    />
-                  </Grid>
-                  <Grid item xs={3}>
-                    <Button
-                      variant="contained"
-                      color="secondary"
-                      className={classes.deleteActivityChoicesButton}
-                      startIcon={<Delete />}
-                      onClick={(e) => {
-                        deleteActivityChoice(index);
-                      }}></Button>
-                  </Grid>
-                </Grid>
-              </Paper>
-            </ListItem>
-          );
-        })}
-      </List>
-    </>
-  );
-};
-
 const PlanPage: React.FC<IPlanPageProps> = (props) => {
   const [geoFeatCollection, setGeoFeatCollection] = useState(null);
 
@@ -310,7 +158,7 @@ const PlanPage: React.FC<IPlanPageProps> = (props) => {
               <Typography className={classes.heading}>Pick Activity Records</Typography>
             </AccordionSummary>
             <AccordionDetails className={classes.activityRecordPicker}>
-              <ActivityDataToCacheChooser />
+              <ActivityDataFilter />
             </AccordionDetails>
           </Accordion>
           <Accordion defaultExpanded={false}>
